ms_decision/requirements.txt
paho-mqtt
json